replaceVanillaLogger option removed Replaced log4jw.xml with vanilla logger Updated to use fabric-installer 0.7.4 server install code Added LICENSE and updated the copyright notices Moved some packages for easier navigation through the code
