+++ gradle.properties
@@ -4,4 +4,4 @@ org.gradle.jvmargs=-Xmx1536M
 artifactid = fabric-wrapper
 group = org.jigsawlabs
 
-version = 0.2
+version = 0.2.2
