+++ src/main/java/org/jigsawlabs/fabricwrapper/WrapperPreLoad.java
@@ -1,6 +1,23 @@
+/*
+ * Copyright (c) 2021 Awakened Redstone
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 package org.jigsawlabs.fabricwrapper;
 
-import org.jigsawlabs.fabricwrapper.notmycode.net.fabricmc.loader.launch.server.InjectingURLClassLoader;
+import org.jigsawlabs.fabricwrapper.loader.launch.server.InjectingURLClassLoader;
 import sun.misc.Unsafe;
 
 import java.lang.invoke.MethodHandle;
