+++ src/main/java/org/jigsawlabs/fabricwrapper/jumploader/launch/PreLaunchDispatcher.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2021  comp500
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial
+ * portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
+ * THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package org.jigsawlabs.fabricwrapper.jumploader.launch;
+
+import org.jigsawlabs.fabricwrapper.jumploader.launch.serviceloading.FileSystemProviderAppender;
+import org.jigsawlabs.fabricwrapper.jumploader.launch.serviceloading.JimfsURLHandlerAppender;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class PreLaunchDispatcher {
+	public interface Handler {
+		void handlePreLaunch(ClassLoader loadingClassloader);
+	}
+
+	private static final List<Handler> HANDLERS = Arrays.asList(
+		new FileSystemProviderAppender(),
+		new JimfsURLHandlerAppender()
+	);
+
+	public static void dispatch(ClassLoader loadingClassloader) {
+		for (Handler handler : HANDLERS) {
+			handler.handlePreLaunch(loadingClassloader);
+		}
+	}
+}
