+++ src/main/java/org/jigsawlabs/fabricwrapper/jumploader/launch/classpath/ClasspathReplacer.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2021-2021  comp500
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial
+ * portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
+ * THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package org.jigsawlabs.fabricwrapper.jumploader.launch.classpath;
+
+import java.io.File;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.List;
+
+public class ClasspathReplacer {
+	private ClasspathReplacer() { }
+
+	public static void replaceClasspath(List<URL> newUrlList) throws URISyntaxException {
+		List<String> newClasspath = new ArrayList<>();
+		for (URL url : newUrlList) {
+			newClasspath.add(Paths.get(url.toURI()).toString());
+		}
+		System.setProperty("java.class.path", String.join(File.pathSeparator, newClasspath));
+	}
+}
