+++ src/main/java/org/jigsawlabs/fabricwrapper/loader/util/UrlConversionException.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2016-2021 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jigsawlabs.fabricwrapper.loader.util;
+
+public class UrlConversionException extends Exception {
+	public UrlConversionException() {
+		super();
+	}
+
+	public UrlConversionException(String s) {
+		super(s);
+	}
+
+	public UrlConversionException(Throwable t) {
+		super(t);
+	}
+
+	public UrlConversionException(String s, Throwable t) {
+		super(s, t);
+	}
+}
