--- src/main/java/org/jigsawlabs/fabricwrapper/notmycode/net/fabricmc/installer/util/InstallerProgress.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jigsawlabs.fabricwrapper.notmycode.net.fabricmc.installer.util;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public interface InstallerProgress {
-
-	InstallerProgress CONSOLE = new InstallerProgress() {
-		private final Logger LOGGER = LogManager.getLogger("FabricWrapper");
-		@Override
-		public void updateProgress(String text) {
-			LOGGER.info(text);
-		}
-
-		@Override
-		public void error(Throwable throwable) {
-			throw new RuntimeException(throwable);
-		}
-	};
-
-	void updateProgress(String text);
-
-	void error(Throwable throwable);
-
-}
