--- src/main/resources/fabric.mod.json
@@ -1,17 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabricwrapper",
-  "name": "Fabric Wrapper",
-  "version": "${version}",
-  "environment": "server",
-  "description": "A wrapper to fix broken Fabric server launchers.",
-  "contact": {
-    "issues": "https://github.com/Awakened-Redstone/FabricWrapper/issues",
-    "sources": "https://github.com/Awakened-Redstone/FabricWrapper"
-  },
-  "license": "Apache-2.0",
-  "icon": "assets/fabricwrapper/icon.png",
-  "authors": [
-    "Awakened Redstone"
-  ]
-}
