--- src/main/resources/libraries.json
@@ -1,8 +1,0 @@
-{
-  "libraries": [
-    {
-      "name": "net.minecrell:terminalconsoleappender:1.2.0",
-      "url": "https://repo.maven.apache.org/maven2/"
-    }
-  ]
-}
