+++ src/main/resources/log4j2.xml
@@ -1,62 +1,28 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="WARN" packages="com.mojang.util,net.minecrell.terminalconsole.util">
+<Configuration status="WARN" packages="com.mojang.util">
 	<Appenders>
-
-		<!--	System out	-->
 		<Console name="SysOut" target="SYSTEM_OUT">
-			<PatternLayout>
-				<LoggerNamePatternSelector defaultPattern="%style{[%d{HH:mm:ss}]}{blue} %style{[%logger{1}]}{cyan} %highlight{[%t/%level]: %msg%n}{FATAL=red bright, ERROR=red, WARN=yellow, INFO=normal, DEBUG=normal, TRACE=normal}">
-					<!-- Dont show the logger name for minecraft classes-->
-					<PatternMatch key="net.minecraft.,com.mojang." pattern="%style{[%d{HH:mm:ss}]}{blue} %style{[Minecraft]}{cyan} %highlight{[%t/%level]: %msg%n}{FATAL=red bright, ERROR=red, WARN=yellow, INFO=normal, DEBUG=normal, TRACE=normal}"/>
-				</LoggerNamePatternSelector>
-			</PatternLayout>
+			<PatternLayout pattern="[%d{HH:mm:ss}] [%t/%level]: %msg%n" />
 		</Console>
-
-		<!--	Vanilla server gui	-->
-		<Queue name="ServerGuiConsole" ignoreExceptions="true">
-			<PatternLayout>
-				<LoggerNamePatternSelector defaultPattern="[%d{HH:mm:ss} %level] [%logger{1}] %msg%n">
-					<!-- Dont show the logger name for minecraft classes-->
-					<PatternMatch key="net.minecraft.,com.mojang." pattern="[%d{HH:mm:ss} %level] %msg%n"/>
-				</LoggerNamePatternSelector>
-			</PatternLayout>
+		<Queue name="ServerGuiConsole">
+			<PatternLayout pattern="[%d{HH:mm:ss} %level]: %msg%n" />
 		</Queue>
-
-		<!--	latest.log same as vanilla	-->
-		<RollingRandomAccessFile name="LatestFile" fileName="logs/latest.log" filePattern="logs/%d{yyyy-MM-dd}-%i.log.gz">
-			<PatternLayout>
-				<LoggerNamePatternSelector defaultPattern="[%d{HH:mm:ss}] [%logger{1}] [%t/%level] %msg%n">
-					<!-- Dont show the logger name for minecraft classes-->
-					<PatternMatch key="net.minecraft.,com.mojang." pattern="[%d{HH:mm:ss}] [Minecraft] [%t/%level] %msg%n"/>
-				</LoggerNamePatternSelector>
-			</PatternLayout>
+		<RollingRandomAccessFile name="File" fileName="logs/latest.log" filePattern="logs/%d{yyyy-MM-dd}-%i.log.gz">
+			<PatternLayout pattern="[%d{HH:mm:ss}] [%t/%level]: %msg%n" />
 			<Policies>
 				<TimeBasedTriggeringPolicy />
 				<OnStartupTriggeringPolicy />
 			</Policies>
 		</RollingRandomAccessFile>
-
-		<!--	Debug log file	-->
-		<RollingRandomAccessFile name="DebugFile" fileName="logs/debug.log" filePattern="logs/debug-%i.log.gz">
-			<PatternLayout pattern="[%d{HH:mm:ss}] [%logger] [%t/%level] %msg%n" />
-
-			<!--	Keep 5 files max	-->
-			<DefaultRolloverStrategy max="5" fileIndex="min"/>
-
-			<Policies>
-				<SizeBasedTriggeringPolicy size="200MB"/>
-				<OnStartupTriggeringPolicy />
-			</Policies>
-
-		</RollingRandomAccessFile>
 	</Appenders>
 	<Loggers>
-		<Logger level="${sys:fabric.log.level:-info}" name="net.minecraft"/>
-		<Root level="all">
-			<AppenderRef ref="DebugFile" level="${sys:fabric.log.debug.level:-debug}"/>
-			<AppenderRef ref="SysOut" level="${sys:fabric.log.level:-info}"/>
-			<AppenderRef ref="LatestFile" level="${sys:fabric.log.level:-info}"/>
-			<AppenderRef ref="ServerGuiConsole" level="${sys:fabric.log.level:-info}"/>
+		<Root level="info">
+			<filters>
+				<MarkerFilter marker="NETWORK_PACKETS" onMatch="DENY" onMismatch="NEUTRAL" />
+			</filters>
+			<AppenderRef ref="SysOut"/>
+			<AppenderRef ref="File"/>
+			<AppenderRef ref="ServerGuiConsole"/>
 		</Root>
 	</Loggers>
 </Configuration>
